backend/requirements.txt
django-filter==23.3
djangorestframework==3.14.0
django-cors-headers==4.3.1